Rename package of restapi and update pom IDs
