+++ .gitignore
@@ -3,7 +3,7 @@
 # compiled output
 /frontend/dist
 /frontend/target
-/backend/target
+/phenopacketlab-restapi/target
 /tmp
 /out-tsc
 # Only exists if Bazel was run
@@ -17,13 +17,14 @@ chrome-profiler-events*.json
 speed-measure-plugin*.json
 
 # IDEs and editors
-/.idea
+.idea/
 .project
 .classpath
 .c9/
 *.launch
 .settings/
 *.sublime-workspace
+*.iml
 
 # IDE - VSCode
 .vscode/*
