+++ frontend/pom.xml
@@ -3,8 +3,8 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <parent>
-    <groupId>org.monarchinitiative.phenopacket</groupId>
-    <artifactId>phenopacket</artifactId>
+    <groupId>org.monarchinitiative.phenopacketlab</groupId>
+    <artifactId>PhenopacketLab</artifactId>
     <version>0.1-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
