+++ phenopacketlab-restapi/pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>org.monarchinitiative.phenopacketlab</groupId>
-        <artifactId>phenopacketlab</artifactId>
+        <artifactId>PhenopacketLab</artifactId>
         <version>0.1-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
@@ -18,31 +18,12 @@
 
     <artifactId>phenopacketlab-restapi</artifactId>
 
-    <dependencies>
-    <dependency>
-        <groupId>${project.groupId}</groupId>
-        <artifactId>frontend</artifactId>
-        <version>${project.version}</version>
-        <type>jar</type>
-    </dependency>
-    </dependencies>
-
     <build>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
                 <version>2.4.3</version>
-                <configuration>
-                    <mainClass>org.monarchinitiative.phenopacketlab.Application</mainClass>
-                </configuration>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>repackage</goal>
-                        </goals>
-                    </execution>
-                </executions>
             </plugin>
 
             <plugin>
