+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.monarchinitiative.phenopacketlab</groupId>
-    <artifactId>phenopacketlab</artifactId>
+    <artifactId>PhenopacketLab</artifactId>
     <packaging>pom</packaging>
     <version>0.1-SNAPSHOT</version>
     <parent>
